Test script for diagnosing start-up problems.
I can be unnecessarily difficult diagnosing problems with the
multi-threading if you can only run MaximaPool within a servlet.
Therefore, I created this test script which can be run from the
command-line to facilitate testing.

It assumes that all the configuration is in a file called
maximapool.conf in the same directorly as the TestMaximaPool.jar
executable is run from.

This tool can also be useful for tuning the pool configuration to get
the quickest possible start-up.